@@ -1,39 +1,44 @@
-const axios = require('axios');
-const SERVER_IP = process.env.REACT_APP_SERVER_IP;
+import axios from "axios";
+import { clearPlayStore } from "~/state/playStore";
+import { loadUser, resetUser, userStore } from "~/state/userStore";
+const SERVER_IP = "https://64vwhnl0nk.execute-api.us-east-1.amazonaws.com/Prod";
 // if not prod server_ip = "" todo
 const register = (email:string, username:string, password:string) => {
-  const params = new URLSearchParams();
-  params.append('email', email);
-  params.append('username', username);
-  params.append('password', password);
-  return axios.post(`${SERVER_IP}/register`, params)
-  .then((response:any) => {
-    return response.data;
+  return axios.post(`${SERVER_IP}/register`, {
+      email,
+      password,
+      username,
+    }, {headers: {'Content-Type': 'application/json'}}
+  )
+  .then((response) => {
+    return response;
   });
 };
 
 const login = (email:string, password:string) => {
-  const params = new URLSearchParams();
-  params.append('email', email);
-  params.append('password', password);
   return axios
-    .post(`${SERVER_IP}/login`, params)
-    .then((response:any) => {
+    .post(`${SERVER_IP}/login`, {
+      email,
+      password
+    }, {headers: {'Content-Type': 'application/json'}})
+    .then((response) => {
       if (response.data.data) {
-        localStorage.setItem("user", JSON.stringify(response.data.data));
+        localStorage.setItem("user", response.data.data);
+        loadUser(response.data.data);
       }
-      return response.data;
+      return response;
     });
 };
 
-const logout = () => {
-  localStorage.removeItem("user");
-  // return axios.post(SERVER_IP + "signout").then((response:any) => {
-  //   return response.data;
-  // });
+const logout = async () => {
+  //todo change from local storage?
+  localStorage.clear();
+  await clearPlayStore();
+  await resetUser();
 };
 const getCurrentUser = () => {
-  return JSON.parse(localStorage.getItem("user") || "{}");
+  // return localStorage.getItem("user") || "";
+  return userStore.data;
 };
 const AuthService = {
   register,
