@@ -1,44 +1,41 @@
-import toast from "react-hot-toast";
-import { PostStat } from "../models/Stat";
-const axios = require('axios');
-const SERVER_IP = process.env.REACT_APP_SERVER_IP;
+import { userStore } from "~/state/userStore";
+import axios from 'axios';
+// const SERVER_IP = process.env.SERVER_IP;
+const SERVER_IP = "https://64vwhnl0nk.execute-api.us-east-1.amazonaws.com/Prod/";
 // if not prod server_ip = "" todo
-const updateStats = (stat:PostStat) => {
-  const params = new URLSearchParams()
-  params.append('stat', JSON.stringify(stat));
+
+const getStats = () => {
+  const auth = localStorage.getItem("user") ? { Authorization: `Bearer ${localStorage.getItem("user")}`} : {};
   return axios
-    .post(`${SERVER_IP}/update-stats`, params)
-    .then((response:any) => {
-      return response.data;
+    .get(`${SERVER_IP}/stats`,
+    {
+      headers: {
+        "Content-Type": "application/json",
+        ...auth
+      }, 
+      params: {
+        page: 1,
+      }
     })
-    .catch((err:any) => {
-      toast.error("Unexpected error...")
+    .then((response) => {
+      return response;
     });
 };
 
-const getStats = (userId:string) => {
-  return axios
-    .get(`${SERVER_IP}/get-stats`,
-    {
-      params: { userId: userId },
-    },
-    {
-      "Content-Type": "application/json"
-    })
-    .then((response:any) => {
-      return response.data;
-    });
-};
 
 const getAllStats = () => {
   return axios
-    .get(`${SERVER_IP}/get-all-stats`,
-    {},
+    .get(`${SERVER_IP}/stats`,
     {
-      "Content-Type": "application/json"
+      headers: {
+        "Content-Type": "application/json"
+      }, 
+      params: {
+        page: 1
+      }
     })
-    .then((response:any) => {
-      return response.data;
+    .then((response) => {
+      return response;
     });
 };
 
@@ -48,16 +45,19 @@ const reportClip = (clip:string | undefined) => {
   params.append('clipSrc', clip);
   return axios
     .post(`${SERVER_IP}/report-clip`, params)
-    .then((response:any) => {
-      return response.data;
+    .then((response) => {
+      return response;
     })
     .catch((err:any) => {
-      toast.error("Unexpected error...")
+      // createToast({
+      //   title: "Unexpected error...",
+      //   duration: 2000,
+      //   placement: "top-end",
+      // })
     });
 };
 
 const UserService = {
-  updateStats,
   getStats,
   getAllStats,
   reportClip
