@@ -1,57 +1,52 @@
-import React, { useCallback, useContext, useState } from "react";
-import {
-  MDBBtn,
-  MDBIcon,
-  MDBNavbar, MDBNavbarBrand, MDBNavbarItem, MDBNavbarNav
-} from "mdb-react-ui-kit";
-import { Link } from "react-router-dom";
 import { LoginModal } from "./LoginModal";
-import { IUser, UserContext } from "../hooks/UseUser";
 import { ProfileModal } from "./ProfileModal";
-import { Alert } from "react-bootstrap";
+import { Component, createSignal } from "solid-js";
+import { Box, AppBar, Toolbar, IconButton, Typography, Button, Link } from "@suid/material";
+import { AiOutlineUser } from "solid-icons/ai";
+import { createStore } from "solid-js/store";
+import AuthService from "~/services/auth.service";
 
+export interface LoginModal {
+  showLogin?: boolean;
+}
+const [state, setState] = createStore<LoginModal>({ showLogin: false});
+export const loginStore = state;
+export const setLoginModal = (show: boolean) => {
+  setState("showLogin", show);
+}
 
-const NavbarPage: React.FC = () => {
-  const [showModal, setShowLoginModal] = useState(false);
-   // eslint-disable-next-line
-   const toggleShowLoginModal = () => setShowLoginModal(!showModal);
-   const [showProfileModal, setShowProfileModal] = useState(false);
-   // eslint-disable-next-line
-  const toggleShowProfileModal = () => setShowProfileModal(!showProfileModal);
-  const { user, updateUser } = useContext<IUser>(UserContext);
-
-  const toggleModal = useCallback(() => {
-    if (user){
-      toggleShowProfileModal()
-    } else {
-      toggleShowLoginModal();
-    }
-  }, [user, toggleShowProfileModal, toggleShowLoginModal]);
+const NavbarPage: Component<any> = ({ updateUser }) => {
+  const [showProfileModal, setShowProfileModal] = createSignal(false);
+  const toggleShowProfileModal = () => setShowProfileModal(!showProfileModal());
   
   return (
       <>
-      <MDBNavbar color="dark" dark expand="md">
-        <MDBNavbarNav fullWidth={false} className='mr-auto mb-2 mb-lg-0'>
-          <MDBNavbarBrand tag="span">
-            <Link to="/">
-              <strong className="logo brand" style={{marginLeft: '20px'}}>MeleeGuessr</strong>
-            </Link>
-          </MDBNavbarBrand>
-        </MDBNavbarNav>
-        <MDBNavbarNav right fullWidth={false} className='mb-2 mb-lg-0'>
-          <MDBNavbarItem style={{marginRight: "20px", alignSelf: 'center'}} active aria-current='page' onClick={toggleModal}>
-            <MDBBtn size="sm" color="light" floating tag={user ? "button" : "a"} >
-              { user ? user.username[0] : <MDBIcon fas icon='user' size='1x'/> }
-            </MDBBtn>
-          </MDBNavbarItem>
-        </MDBNavbarNav>
-      </MDBNavbar>
-      <Alert variant='danger'>
-				Unfortunately the servers are too expensive for me to maintain by myself so the website will be shutting down on 5/15/22, only 7 days after launch. Thanks for playing! - Noah
-			</Alert>
-			
-      <ProfileModal showModal={showProfileModal} setShowModal={setShowProfileModal} toggleShowModal={toggleShowProfileModal} updateUser={updateUser} />
-      <LoginModal showModal={showModal} setShowModal={setShowLoginModal} toggleShowModal={toggleShowLoginModal} updateUser={updateUser} />
+        <Box sx={{ flexGrow: 1 }}>
+        <AppBar position="fixed">
+          <Toolbar variant="dense">
+            <IconButton
+              size="large"
+              edge="start"
+              color="inherit"
+              aria-label="menu"
+              sx={{ mr: 2 }}
+            >
+            </IconButton>
+            <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+              <Link href="/" sx={{display: 'flex', 'alignItems': 'center'}}><h1 class="navbar-brand logo">MeleeGuessr</h1></Link>
+            </Typography>
+            
+			      { AuthService.getCurrentUser() ?
+              <Button  onClick={() => setShowProfileModal(!showProfileModal())} >
+                <AiOutlineUser size={24} />
+              </Button> :
+              <Button onClick={() => setState("showLogin", !state.showLogin)} color="inherit">Login</Button>
+            }
+            </Toolbar>
+          </AppBar>
+        </Box>
+        <ProfileModal showModal={showProfileModal} setShowModal={setShowProfileModal} toggleShowModal={toggleShowProfileModal} updateUser={updateUser} />
+        <LoginModal updateUser={updateUser} />
     </>
     );
   }
