@@ -6,10 +6,25 @@ type conditionalCharacter = {
 
 export type Choice = {
   label: string;
+  characters: Character[],
   color?: string;
   aliases?: string[];
-  characters?: Character[];
-  conditionalCharacters?: conditionalCharacter;
+  conditionalCharacters?: conditionalCharacter,
   flag?: string;
   score?: number;
+  controllerType?: ControllerType;
+}
+
+export interface ControllerType {
+  type: {
+    rectangle?: "b0xx" | "smashbox" | "frame1" | "other" | boolean;
+    phob?: boolean;
+    oEMModded?: boolean;
+    goomwave?: boolean;
+  };
+  mods?: {
+    notches?: boolean;
+    zJump?: boolean;
+    snapback?: boolean;
+  };
 }
