@@ -1,33 +1,19 @@
-import { MDBBtn } from "mdb-react-ui-kit";
-import { PlayData } from "./Play";
+import { Button, Container, Grid, Typography } from '@suid/material';
+import { useNavigate } from '@solidjs/router';
 
-interface ResultProps {
-  HS: boolean;
-  score: number;
-  playData: PlayData[];
-  reset: () => void;
-}
-
-export const Result: React.FC<ResultProps> = ({ HS, score, playData, reset }) => {
+export const Result = () => {
+  const navigate = useNavigate();
+  const HS = '9';
   return (
     <>
-      <div className="full-page d-flex justify-content-center align-items-center">
-				<div className="row justify-content-center" style={{textAlign: 'center'}}>
-			    {/* <div className="row" style={{textAlign: "center"}}> */}
-          { HS && <h1>High Score!</h1> }
-          <h2 className="white-text">Final: {score}%</h2>
-            <MDBBtn size="lg" color="success" onClick={reset}>Retry?</MDBBtn>
-            {/* <MDBBtn className="mt-2" color="danger" onClick={() => setStop(true)}>Stop</MDBBtn> */}
-          {/* </div> */}
-				</div>
-			</div>
-      {/* { playData.map((data, i) => {
-        return (
-          <div key={i} className="row w-100 secondary-text" style={{textAlign: "center"}}>
-            <h3>#{i+1}. {data.wasCorrect ? "Correct" : "Incorrect"}, You chose {data.choice}</h3>
-          </div> 
-        )
-      }) } */}
+      <Grid container minHeight="90vh" alignItems='center' justifyContent="center" class="nav-m">
+        <Container maxWidth="md">
+          <Grid sx={{width: "100%", mt: 5, mb: 5, textAlign: 'center'}} xs={12} justifyContent="center">
+            { HS && <Typography variant='h2'>High Score! {HS}</Typography> }
+            <Button color="secondary" variant="contained" onClick={() => navigate('/play')}>Retry?</Button>
+          </Grid>
+        </Container>
+      </Grid>
     </>
   )
 }
